+++ Pandora/Paradigm/Basis/Jack.hs
@@ -5,7 +5,7 @@ import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)), comap)
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Alternative (Alternative ((<+>)))
-import Pandora.Pattern.Functor.Avoidable (Avoidable (idle))
+import Pandora.Pattern.Functor.Avoidable (Avoidable (empty))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>), traverse))
 import Pandora.Pattern.Functor.Distributive (Distributive ((>>-), distribute))
@@ -28,7 +28,7 @@ instance Alternative t => Alternative (Jack t) where
 	Other x <+> Other y = Other (x <+> y)
 
 instance Avoidable t => Avoidable (Jack t) where
-	idle = Other idle
+	empty = Other empty
 
 instance Extractable t => Extractable (Jack t) where
 	extract (It x) = x