+++ Pandora/Paradigm/Basis/Jet.hs
@@ -1,7 +1,7 @@
 module Pandora.Paradigm.Basis.Jet (Jet (..)) where
 
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)), (<$$>))
-import Pandora.Pattern.Functor.Avoidable (Avoidable (idle))
+import Pandora.Pattern.Functor.Avoidable (Avoidable (empty))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Extractable (Extractable (extract))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
@@ -18,7 +18,7 @@ instance Traversable t => Traversable (Jet t) where
 	a :- as ->> f = (:-) <$> f a <*> as ->>> f
 
 instance (forall t' . Avoidable t') => Pointable (Jet t) where
-	point x = x :- idle
+	point x = x :- empty
 
 instance Covariant t => Extractable (Jet t) where
 	extract (x :- _) = x