+++ Pandora/Pattern/Functor/Avoidable.hs
@@ -4,9 +4,9 @@ import Pandora.Pattern.Functor.Alternative (Alternative)
 
 {- |
 > When providing a new instance, you should ensure it satisfies the two laws:
-> * Left absorption: x <+> idle ≡ x
-> * Right absorption: idle <+> x ≡ x
+> * Left absorption: x <+> empty ≡ x
+> * Right absorption: empty <+> x ≡ x
 -}
 
 class Alternative t => Avoidable t where
-	idle :: t a
+	empty :: t a